price_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
casabourse>=0.1.5
python-dotenv==1.0.1
requests>=2.31.0
pytz>=2024.1
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.11
